+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 1. remove `beforeInitiation` & `afterInitiation` from `FormeFieldState` , use `initState` instead
 2. remove `updateFieldValueInDidUpdateWidget` , use `didUpdateValue` to tell user to whether field value needs update or not after didUpdateWidget
 3. add `FormeFieldValueUpdater` on `FormeField` , used to update value when `didUpdateValue` return true
+4. initialValue on `Forme` has higher priority than `FormeField`'s initialValue
 
 ## 3.5.2